add checks for flaking test
