+++ containers/tefca-viewer/e2e/constants.ts
@@ -1,0 +1,5 @@
+import { hyperUnluckyPatient } from "@/app/constants";
+
+export const TEST_PATIENT = hyperUnluckyPatient;
+export const TEST_PATIENT_NAME =
+  hyperUnluckyPatient.FirstName + " A. " + hyperUnluckyPatient.LastName;
